main,refactor: split subparser.h into subparser.h and subparser_p.h
