+++ win32/ctags_vs2013.vcxproj.filters
@@ -638,6 +638,9 @@
     <ClInclude Include="..\main\subparser.h">
       <Filter>Header Files</Filter>
     </ClInclude>
+    <ClInclude Include="..\main\subparser_p.h">
+      <Filter>Header Files</Filter>
+    </ClInclude>
     <ClInclude Include="..\main\tokeninfo.h">
       <Filter>Header Files</Filter>
     </ClInclude>
